Moved get_track_info_and_features into own file along with authentication bypass
